--- src/main/java/search_engine/decoders/Decoder.java
@@ -1,5 +1,0 @@
-package search_engine.decoders;
-
-public interface Decoder {
-    Query decode(String query);
-}
