+++ src/main/java/search_engine/queryDecoders/QueryDecoder.java
@@ -1,0 +1,5 @@
+package search_engine.queryDecoders;
+
+public interface QueryDecoder {
+    Query decode(String query);
+}
